// generated/models/ApiCollectTdhRanking.ts
export class ApiCollectTdhRanking {
    'ranking_id': string;
    'account': ApiCollectAccount;
    'snapshot_block': number;
    'snapshot_timestamp': string;
    'acquisition_timestamp': string;
    'evaluated_at': string;
    'horizon_days': number;
    'rules_version': string;
    'candidate_count': number;
    'evaluated_count': number;
    'optimality': ApiCollectTdhRankingOptimalityEnum;
    'candidate_scope': ApiCollectTdhRankingCandidateScopeEnum;
    'ranked': Array<ApiCollectTdhRankedCandidate>;
    'excluded': Array<ApiCollectTdhExcludedCandidate>;
    'assumptions': Array<string>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "ranking_id",
            "baseName": "ranking_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "account",
            "baseName": "account",
            "type": "ApiCollectAccount",
            "format": ""
        },
        {
            "name": "snapshot_block",
            "baseName": "snapshot_block",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "snapshot_timestamp",
            "baseName": "snapshot_timestamp",
            "type": "string",
            "format": ""
        },
        {
            "name": "acquisition_timestamp",
            "baseName": "acquisition_timestamp",
            "type": "string",
            "format": ""
        },
        {
            "name": "evaluated_at",
            "baseName": "evaluated_at",
            "type": "string",
            "format": ""
        },
        {
            "name": "horizon_days",
            "baseName": "horizon_days",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "rules_version",
            "baseName": "rules_version",
            "type": "string",
            "format": ""
        },
        {
            "name": "candidate_count",
            "baseName": "candidate_count",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "evaluated_count",
            "baseName": "evaluated_count",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "optimality",
            "baseName": "optimality",
            "type": "ApiCollectTdhRankingOptimalityEnum",
            "format": ""
        },
        {
            "name": "candidate_scope",
            "baseName": "candidate_scope",
            "type": "ApiCollectTdhRankingCandidateScopeEnum",
            "format": ""
        },
        {
            "name": "ranked",
            "baseName": "ranked",
            "type": "Array<ApiCollectTdhRankedCandidate>",
            "format": ""
        },
        {
            "name": "excluded",
            "baseName": "excluded",
            "type": "Array<ApiCollectTdhExcludedCandidate>",
            "format": ""
        },
        {
            "name": "assumptions",
            "baseName": "assumptions",
            "type": "Array<string>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiCollectTdhRanking.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiCollectTdhRankingOptimalityEnum {
    BestFound = 'best_found'
}
export enum ApiCollectTdhRankingCandidateScopeEnum {
    SuppliedVerifiedQuotes = 'supplied_verified_quotes'
}